Project contacts now export and import
Deleted the ContactList class and moved all needed functionality into
the ContactManager class. 

Deleted the ContactListTest class.

Added a timestamp to the Contact class to handle how contacts are merged
when imported.

Added a ContactManagerListener to be used for listening to updates from
the ContactManager.

Fixed tests that were broken from changes to the Contact class.